Add Bootstrap base UI for Reviews, add to 2 sidebar's item of user info to navigate to specified reviews , fix authorization of review

+++ src/main/java/gr/careplus4/entities/ReviewDetail.java
@@ -7,6 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -33,5 +34,5 @@ public class ReviewDetail implements Serializable {
     @Column(name = "Rating", nullable = false)
     @DecimalMin("0.0")
     @DecimalMax("5.0")
-    private float rating;
+    private BigDecimal rating;
 }
